Update install-on-kubernetes.md
I propose to expand on the topic of variables that are taken from a Kubernetes configmap or secret, personally it was very difficult for me to find this information because it does not exist anywhere, I had to read the classes in charge of this operation, I also think that the syntax is not very clear, so I made a small modification. I remain of you.